Factor out code into its own module
